Cleanup sys.version_info usage
- Remove python version checks that are now unnecessary.
- Centralize all python version checks on a global defined in
`dask.compatibility` to make grepping and removing them later much
easier.
