+++ dask/bag/tests/test_text.py
@@ -1,12 +1,11 @@
 from __future__ import print_function, division, absolute_import
 
-import sys
-
 import pytest
 from toolz import partial
 
 import dask
 from dask import compute
+from dask.compatibility import PY_VERSION
 from dask.utils import filetexts
 from dask.bytes import utils
 from dask.bag.text import read_text
@@ -41,7 +40,7 @@ fmt_bs_enc = [(fmt, bs, encoding) for fmt, bs in fmt_bs for encoding in encoding
 
 @pytest.mark.parametrize("fmt,bs,encoding", fmt_bs_enc)
 def test_read_text(fmt, bs, encoding):
-    if fmt == "zip" and sys.version_info.minor == 5:
+    if fmt == "zip" and PY_VERSION < "3.6":
         pytest.skip("zipfile is read-only on py35")
     if fmt not in utils.compress:
         pytest.skip("compress function not provided for %s" % fmt)
